// src/types/job.d.ts
export type Job =
  | "NONE"
  | "GLA"
  | "PGL"
  | "MRD"
  | "LNC"
  | "ARC"
  | "CNJ"
  | "THM"
  | "CRP"
  | "BSM"
  | "ARM"
  | "GSM"
  | "LTW"
  | "WVR"
  | "ALC"
  | "CUL"
  | "MIN"
  | "BTN"
  | "FSH"
  | "PLD"
  | "MNK"
  | "WAR"
  | "DRG"
  | "BRD"
  | "WHM"
  | "BLM"
  | "ACN"
  | "SMN"
  | "SCH"
  | "ROG"
  | "NIN"
  | "MCH"
  | "DRK"
  | "AST"
  | "SAM"
  | "RDM"
  | "BLU"
  | "GNB"
  | "DNC"
  | "RPR"
  | "SGE"
  | "VPR"
  | "PCT";

export type Role = "tank" | "healer" | "dps" | "crafter" | "gatherer" | "none";

export type FFIcon =
  | 'NONE'
  | 'Gladiator'
  | 'Pugilist'
  | 'Marauder'
  | 'Lancer'
  | 'Archer'
  | 'Conjurer'
  | 'Thaumaturge'
  | 'Carpenter'
  | 'Blacksmith'
  | 'Armorer'
  | 'Goldsmith'
  | 'Leatherworker'
  | 'Weaver'
  | 'Alchemist'
  | 'Culinarian'
  | 'Miner'
  | 'Botanist'
  | 'Fisher'
  | 'Paladin'
  | 'Monk'
  | 'Warrior'
  | 'Dragoon'
  | 'Bard'
  | 'WhiteMage'
  | 'BlackMage'
  | 'Arcanist'
  | 'Summoner'
  | 'Scholar'
  | 'Rogue'
  | 'Ninja'
  | 'Machinist'
  | 'DarkKnight'
  | 'Astrologian'
  | 'Samurai'
  | 'RedMage'
  | 'BlueMage'
  | 'Gunbreaker'
  | 'Dancer'
  | 'Reaper'
  | 'Sage'
  | 'Viper'
  | 'Pictomancer'